Move tax TextUtil to real TextUtil Fix some messaging code
