+++ src/main/java/com/massivecraft/factions/tax/format/TimeDiffUtil.java
@@ -12,6 +12,7 @@ import java.util.regex.Pattern;
 
 import com.avaje.ebean.TxType;
 import com.massivecraft.factions.P;
+import com.massivecraft.factions.zcore.util.TextUtil;
 
 public class TimeDiffUtil
 {
